+++ src/settings/AppSettings.ts
@@ -6,10 +6,12 @@
  */
 import { AppThemePreference } from "./AppThemePreference";
 import { DefaultNoteColorPreference } from "./noteColorPreference";
+import { NoteSortOrder } from "./NoteSortOrder";
 
 export type AppSettings = {
   showWelcomeScreenOnLaunch: boolean;
   keepNotesMainWindowOnTop: boolean;
+  notesSortOrder: NoteSortOrder;
   theme: AppThemePreference;
   defaultNoteColor: DefaultNoteColorPreference;
 };