+++ src/settings/NoteSortOrder.ts
@@ -1,0 +1,13 @@
+/**
+ * Copyright (c) 2026 SolisWare.
+ *
+ * All rights reserved. Licensed under the MIT license.
+ * See the LICENSE.txt file in the project root directory for details.
+ */
+export enum NoteSortOrder {
+  DATE_CREATED_ASC = "dateCreatedAsc",
+  DATE_CREATED_DESC = "dateCreatedDesc",
+  LAST_MODIFIED = "lastModified",
+  TITLE_ASC = "titleAsc",
+  TITLE_DESC = "titleDesc"
+}