+++ src/views/MainWindow/pages/Home.tsx
@@ -6,10 +6,12 @@
  */
 import NoteList from "../../../components/NoteList";
 import { NoteType } from "../../../models/NoteType";
+import { NoteSortOrder } from "../../../settings/NoteSortOrder";
 import { SystemTheme } from "../../../theme/SystemTheme";
 
 type HomeProps = {
   notes: NoteType[];
+  notesSortOrder: NoteSortOrder;
   theme: SystemTheme;
   handleDeleteNoteButton: (noteId: string) => void;
 }
@@ -17,7 +19,12 @@ type HomeProps = {
 function Home(props: HomeProps) {
   return (
     <div>
-      <NoteList theme={props.theme} notes={props.notes} handleDeleteNoteButton={props.handleDeleteNoteButton} />
+      <NoteList
+        theme={props.theme}
+        notes={props.notes}
+        notesSortOrder={props.notesSortOrder}
+        handleDeleteNoteButton={props.handleDeleteNoteButton}
+      />
     </div>
   );
 }